NNUE Implementation: add some code (part 2) - experiments, fixes, adjustments

+++ Sources/Bitboard/src/bagaturchess/bitboard/impl1/BoardImpl.java
@@ -681,7 +681,7 @@ public class BoardImpl implements IBitBoard {
 	
 	
 	@Override
-	public double[] getNNUEInputs() {
+	public float[] getNNUEInputs() {
 		
 		return nnue_input.getInputs();
 	}
